requirements.txt
coverage==7.12.0
django==5.2.8
dj-database-url==3.1.0
flake8==7.3.0
pytest==8.3.3
pytest-cov==7.0.0
pytest-django==4.9.0
sentry-sdk==2.19.2
python-dotenv==1.0.1